refactor: user api adapterクラスに処理を切り出し
