+++ modules/data/src/main/java/net/pantasystem/milktea/data/infrastructure/user/UserRepositoryImpl.kt
@@ -1,8 +1,10 @@
 package net.pantasystem.milktea.data.infrastructure.user
 
 
-import kotlinx.coroutines.*
-import net.pantasystem.milktea.api.misskey.users.*
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.withContext
+import net.pantasystem.milktea.api.misskey.users.RequestUser
+import net.pantasystem.milktea.api.misskey.users.from
 import net.pantasystem.milktea.common.Logger
 import net.pantasystem.milktea.common.runCancellableCatching
 import net.pantasystem.milktea.common.throwIfHasError
@@ -12,7 +14,6 @@ import net.pantasystem.milktea.data.api.misskey.MisskeyAPIProvider
 import net.pantasystem.milktea.data.converters.UserDTOEntityConverter
 import net.pantasystem.milktea.data.infrastructure.notes.reaction.impl.history.ReactionHistoryDao
 import net.pantasystem.milktea.data.infrastructure.toUserRelated
-import net.pantasystem.milktea.model.account.Account
 import net.pantasystem.milktea.model.account.AccountRepository
 import net.pantasystem.milktea.model.drive.FilePropertyDataSource
 import net.pantasystem.milktea.model.user.User
@@ -213,49 +214,11 @@ internal class UserRepositoryImpl @Inject constructor(
     override suspend fun syncIn(userIds: List<User.Id>): Result<List<User.Id>> {
         return runCancellableCatching {
             withContext(ioDispatcher) {
-                val accountIds = userIds.map { it.accountId }.distinct()
-                if (accountIds.isEmpty()) {
-                    return@withContext emptyList()
+                val users = userApiAdapter.showUsers(userIds, true)
+                userDataSource.addAll(users)
+                users.map {
+                    it.id
                 }
-                coroutineScope {
-                    accountIds.map { accountId ->
-                        async {
-                            val account = accountRepository.get(accountId).getOrThrow()
-                            when(account.instanceType) {
-                                Account.InstanceType.MISSKEY, Account.InstanceType.FIREFISH -> {
-                                    val users = misskeyAPIProvider.get(account)
-                                        .showUsers(
-                                            RequestUser(
-                                                i = account.token,
-                                                userIds = userIds.filter { it.accountId == accountId }.map { it.id },
-                                                detail = true
-                                            )
-                                        ).throwIfHasError()
-                                        .body()!!.map {
-                                            userDTOEntityConverter.convert(account, it, true)
-                                        }
-                                    userDataSource.addAll(users)
-                                    users.map { it.id }
-                                }
-                                Account.InstanceType.MASTODON, Account.InstanceType.PLEROMA -> {
-                                    val users = userIds.filter { it.accountId == accountId }.map { it.id }.map {
-                                        async {
-                                            requireNotNull(
-                                                mastodonAPIProvider.get(account)
-                                                    .getAccount(it)
-                                                    .throwIfHasError()
-                                                    .body()
-                                            ).toModel(account)
-                                        }
-                                    }.awaitAll()
-                                    userDataSource.addAll(users)
-                                    users.map { it.id }
-                                }
-                            }
-
-                        }
-                    }.awaitAll()
-                }.flatten()
             }
         }
     }
